requirements.txt
# Dependencias para Visualizador de Eventos Geotécnicos
streamlit
pandas
plotly
folium
streamlit-folium
numpy
openpyxl
xlsxwriter
pillow
ezdxf
numpy-stl
